[SCALA-97] Renamed the package with the examples of the tell pattern
